+++ src/Axle/Startup.cs
@@ -125,6 +125,9 @@ namespace Axle
                         options.IntrospectionDiscoveryPolicy.ValidateIssuerName = validateIssuerName;
 
                         options.TokenRetriever = BearerTokenRetriever.FromHeaderAndQueryString;
+
+                        options.EnableCaching = this.configuration.GetValue("IntrospectionCache:Enabled", true);
+                        options.CacheDuration = TimeSpan.FromSeconds(this.configuration.GetValue("IntrospectionCache:DurationInSeconds", 600));
                     });
 
             var connectionRepository = new InMemoryRepository<string, HubCallerContext>();
@@ -200,7 +203,11 @@ namespace Axle
             services.AddSingleton<IAuthorizationHandler, MobileClientAndAccountOwnerHandler>();
             services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
             services.AddSingleton<IAccountsCache, AccountsCache>();
+
             services.AddMemoryCache(o => o.ExpirationScanFrequency = TimeSpan.FromMinutes(1));
+            services.AddDistributedMemoryCache(
+                options => options.ExpirationScanFrequency = TimeSpan.FromSeconds(
+                    this.configuration.GetValue("IntrospectionCache:ExpirationScanFrequencyInSeconds", 60)));
         }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)