+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Unreleased
 
 * LT-1240: Update Licenses
+* LT-1264: Enabled introspection cache
+
+### Configuration changes
+
+  - Added configuration for reference token introspection cache
+
+```json
+  "IntrospectionCache": {
+    "Enabled": true,
+    "DurationInSeconds": 600,
+    "ExpirationScanFrequencyInSeconds": 60
+  }
+```
 
 ## 2.12.0 (March 27, 2019)